Add intial information on defining a variant.
This addition is a guide for board makers to use when defining
meshtastic variants. Currently it has information about settings for
GPS, Audio, Buzzers, and Keyboards.

Generally, variant.h is a bit difficult to follow right now - there are
more than 500 variables. This documentation work is a companion to
changes like meshtastic/firmware#4421 that aim to clean up the
proliferation of duplicate variables.